refactor(diag): Move lints to diagnostics
Now that we more formally have a diagnostic in here (#16976),
let's make the broader intent more clear.

See #16975 for background on where this is going.
