+++ src/cargo/core/workspace.rs
@@ -21,25 +21,25 @@ use crate::core::{
     PatchLocation,
 };
 use crate::core::{EitherManifest, Package, SourceId, VirtualManifest};
-use crate::lints::rules::blanket_hint_mostly_unused;
-use crate::lints::rules::check_im_a_teapot;
-use crate::lints::rules::implicit_minimum_version_req_pkg;
-use crate::lints::rules::implicit_minimum_version_req_ws;
-use crate::lints::rules::missing_lints_features;
-use crate::lints::rules::missing_lints_inheritance;
-use crate::lints::rules::non_kebab_case_bins;
-use crate::lints::rules::non_kebab_case_features;
-use crate::lints::rules::non_kebab_case_packages;
-use crate::lints::rules::non_snake_case_features;
-use crate::lints::rules::non_snake_case_packages;
-use crate::lints::rules::redundant_homepage;
-use crate::lints::rules::redundant_readme;
-use crate::lints::rules::text_direction_codepoint_in_comment;
-use crate::lints::rules::text_direction_codepoint_in_literal;
-use crate::lints::rules::unknown_lints;
-use crate::lints::rules::unused_build_dependencies_no_build_rs;
-use crate::lints::rules::unused_workspace_dependencies;
-use crate::lints::rules::unused_workspace_package_fields;
+use crate::diagnostics::rules::blanket_hint_mostly_unused;
+use crate::diagnostics::rules::check_im_a_teapot;
+use crate::diagnostics::rules::implicit_minimum_version_req_pkg;
+use crate::diagnostics::rules::implicit_minimum_version_req_ws;
+use crate::diagnostics::rules::missing_lints_features;
+use crate::diagnostics::rules::missing_lints_inheritance;
+use crate::diagnostics::rules::non_kebab_case_bins;
+use crate::diagnostics::rules::non_kebab_case_features;
+use crate::diagnostics::rules::non_kebab_case_packages;
+use crate::diagnostics::rules::non_snake_case_features;
+use crate::diagnostics::rules::non_snake_case_packages;
+use crate::diagnostics::rules::redundant_homepage;
+use crate::diagnostics::rules::redundant_readme;
+use crate::diagnostics::rules::text_direction_codepoint_in_comment;
+use crate::diagnostics::rules::text_direction_codepoint_in_literal;
+use crate::diagnostics::rules::unknown_lints;
+use crate::diagnostics::rules::unused_build_dependencies_no_build_rs;
+use crate::diagnostics::rules::unused_workspace_dependencies;
+use crate::diagnostics::rules::unused_workspace_package_fields;
 use crate::ops;
 use crate::ops::lockfile::LOCKFILE_NAME;
 use crate::sources::{CRATES_IO_INDEX, CRATES_IO_REGISTRY, PathSource, SourceConfigMap};
