+++ src/cargo/diagnostics/report.rs
