+++ src/cargo/diagnostics/rules/missing_lints_inheritance.rs
@@ -13,9 +13,9 @@ use crate::CargoResult;
 use crate::GlobalContext;
 use crate::core::Package;
 use crate::core::Workspace;
-use crate::lints::Lint;
-use crate::lints::LintLevel;
-use crate::lints::rel_cwd_manifest_path;
+use crate::diagnostics::Lint;
+use crate::diagnostics::LintLevel;
+use crate::diagnostics::rel_cwd_manifest_path;
 
 pub static LINT: &Lint = &Lint {
     name: "missing_lints_inheritance",
