+++ src/cargo/diagnostics/rules/redundant_homepage.rs
@@ -1,7 +1,6 @@
 use std::path::Path;
 
 use cargo_util_schemas::manifest::InheritableField;
-use cargo_util_schemas::manifest::StringOrBool;
 use cargo_util_schemas::manifest::TomlToolLints;
 use cargo_util_terminal::report::AnnotationKind;
 use cargo_util_terminal::report::Group;
@@ -15,16 +14,15 @@ use super::STYLE;
 use crate::CargoResult;
 use crate::GlobalContext;
 use crate::core::Package;
-use crate::lints::Lint;
-use crate::lints::LintLevel;
-use crate::lints::LintLevelSource;
-use crate::lints::get_key_value_span;
-use crate::lints::rel_cwd_manifest_path;
-use crate::util::toml::DEFAULT_README_FILES;
+use crate::diagnostics::Lint;
+use crate::diagnostics::LintLevel;
+use crate::diagnostics::LintLevelSource;
+use crate::diagnostics::get_key_value_span;
+use crate::diagnostics::rel_cwd_manifest_path;
 
 pub static LINT: &Lint = &Lint {
-    name: "redundant_readme",
-    desc: "explicit `package.readme` can be inferred",
+    name: "redundant_homepage",
+    desc: "`package.homepage` is redundant with another manifest field",
     primary_group: &STYLE,
     msrv: Some(super::CARGO_LINTS_MSRV),
     feature_gate: None,
@@ -32,24 +30,23 @@ pub static LINT: &Lint = &Lint {
         r#"
 ### What it does
 
-Checks for `package.readme` fields that can be inferred.
+Checks if the value of `package.homepage` is already covered by another field.
 
-See also [`package.readme` reference documentation](manifest.md#the-readme-field).
+See also [`package.homepage` reference documentation](manifest.md#the-homepage-field).
 
 ### Why it is bad
 
-Adds boilerplate.
+When package browsers render each link, a redundant link adds visual noise.
 
 ### Drawbacks
-
-It might not be obvious if they named their file correctly.
 
 ### Example
 
 ```toml
 [package]
 name = "foo"
-readme = "README.md"
+homepage = "https://github.com/rust-lang/cargo/"
+repository = "https://github.com/rust-lang/cargo/"
 ```
 
 Should be written as:
@@ -57,13 +54,14 @@ Should be written as:
 ```toml
 [package]
 name = "foo"
+repository = "https://github.com/rust-lang/cargo/"
 ```
 "#,
     ),
 };
 
 #[instrument(skip_all)]
-pub fn redundant_readme(
+pub fn redundant_homepage(
     pkg: &Package,
     manifest_path: &Path,
     cargo_lints: &TomlToolLints,
@@ -95,26 +93,24 @@ fn lint_package(
 ) -> CargoResult<()> {
     let manifest = pkg.manifest();
 
-    // Must check `original_toml`, before any inferring happened
-    let Some(original_toml) = manifest.original_toml() else {
+    let Some(normalized_pkg) = &manifest.normalized_toml().package else {
         return Ok(());
     };
-    let Some(original_pkg) = &original_toml.package else {
-        return Ok(());
-    };
-    let Some(readme) = &original_pkg.readme else {
+    let Some(InheritableField::Value(homepage)) = &normalized_pkg.homepage else {
         return Ok(());
     };
 
-    let InheritableField::Value(StringOrBool::String(readme)) = readme else {
-        // Not checking inheritance because at most one package can be identified from the lint and
-        // consistency of inheritance is likely best.
+    let other_field = if let Some(InheritableField::Value(repository)) = &normalized_pkg.repository
+        && repository == homepage
+    {
+        "repository"
+    } else if let Some(InheritableField::Value(documentation)) = &normalized_pkg.documentation
+        && documentation == homepage
+    {
+        "documentation"
+    } else {
         return Ok(());
     };
-
-    if !DEFAULT_README_FILES.contains(&readme.as_str()) {
-        return Ok(());
-    }
 
     let document = manifest.document();
     let contents = manifest.contents();
@@ -124,14 +120,15 @@ fn lint_package(
     let mut primary = Group::with_title(level.primary_title(LINT.desc));
     if let Some(document) = document
         && let Some(contents) = contents
-        && let Some(span) = get_key_value_span(document, &["package", "readme"])
+        && let Some(span) = get_key_value_span(document, &["package", "homepage"])
     {
-        let span = span.key.start..span.value.end;
-        primary = primary.element(
-            Snippet::source(contents)
-                .path(manifest_path)
-                .annotation(AnnotationKind::Primary.span(span)),
-        );
+        let mut snippet = Snippet::source(contents)
+            .path(manifest_path)
+            .annotation(AnnotationKind::Primary.span(span.value));
+        if let Some(span) = get_key_value_span(document, &["package", other_field]) {
+            snippet = snippet.annotation(AnnotationKind::Context.span(span.value));
+        }
+        primary = primary.element(snippet);
     } else {
         primary = primary.element(Origin::path(manifest_path));
     }
@@ -139,17 +136,17 @@ fn lint_package(
     let mut report = vec![primary];
     if let Some(document) = document
         && let Some(contents) = contents
-        && let Some(span) = get_key_value_span(document, &["package", "readme"])
+        && let Some(span) = get_key_value_span(document, &["package", "homepage"])
     {
         let span = if let Some(workspace_span) =
-            get_key_value_span(document, &["package", "readme", "workspace"])
+            get_key_value_span(document, &["package", "homepage", "workspace"])
         {
             span.key.start..workspace_span.value.end
         } else {
             span.key.start..span.value.end
         };
         let mut help =
-            Group::with_title(Level::HELP.secondary_title("consider removing `package.readme`"));
+            Group::with_title(Level::HELP.secondary_title("consider removing `package.homepage`"));
         help = help.element(
             Snippet::source(contents)
                 .path(manifest_path)
