+++ src/cargo/diagnostics/rules/text_direction_codepoint_in_literal.rs
@@ -4,6 +4,7 @@ use cargo_util_schemas::manifest::TomlToolLints;
 use cargo_util_terminal::report::AnnotationKind;
 use cargo_util_terminal::report::Group;
 use cargo_util_terminal::report::Level;
+use cargo_util_terminal::report::Patch;
 use cargo_util_terminal::report::Snippet;
 use toml_parser::Source;
 use toml_parser::Span;
@@ -17,27 +18,27 @@ use super::CORRECTNESS;
 use crate::CargoResult;
 use crate::GlobalContext;
 use crate::core::MaybePackage;
-use crate::lints::Lint;
-use crate::lints::LintLevel;
-use crate::lints::ManifestFor;
-use crate::lints::rel_cwd_manifest_path;
+use crate::diagnostics::Lint;
+use crate::diagnostics::LintLevel;
+use crate::diagnostics::ManifestFor;
+use crate::diagnostics::rel_cwd_manifest_path;
 
 pub static LINT: &Lint = &Lint {
-    name: "text_direction_codepoint_in_comment",
-    desc: "unicode codepoint changing visible direction of text present in comment",
+    name: "text_direction_codepoint_in_literal",
+    desc: "unicode codepoint changing visible direction of text present in literal",
     primary_group: &CORRECTNESS,
     msrv: Some(super::CARGO_LINTS_MSRV),
     feature_gate: None,
     docs: Some(
         r#"
 ### What it does
-Detects Unicode codepoints in manifest comments that change the visual representation of text on screen
+Detects Unicode codepoints in literals in manifests that change the visual representation of text on screen
 in a way that does not correspond to their on memory representation.
 
 ### Why it is bad
 Unicode allows changing the visual flow of text on screen
 in order to support scripts that are written right-to-left,
-but a specially crafted comment can make code that will be compiled appear to be part of a comment,
+but a specially crafted literal can make code that will be compiled appear to be part of a literal,
 depending on the software used to read the code.
 To avoid potential problems or confusion,
 such as in CVE-2021-42574,
@@ -47,7 +48,7 @@ by default we deny their use.
 };
 
 #[instrument(skip_all)]
-pub fn text_direction_codepoint_in_comment(
+pub fn text_direction_codepoint_in_literal(
     manifest: ManifestFor<'_>,
     manifest_path: &Path,
     cargo_lints: &TomlToolLints,
@@ -79,7 +80,7 @@ pub fn text_direction_codepoint_in_comment(
         .filter(|(_i, c)| {
             UNICODE_BIDI_CODEPOINTS
                 .iter()
-                .any(|(bidi, _name)| c == bidi)
+                .any(|(bidi, _, _name)| c == bidi)
         })
         .map(|(i, c)| (i, i + c.len_utf8()))
         .collect::<Vec<_>>();
@@ -87,7 +88,8 @@ pub fn text_direction_codepoint_in_comment(
         return Ok(());
     }
 
-    let events = bidi_events(contents, &bidi_spans);
+    let toml_source = Source::new(contents);
+    let events = bidi_events(&toml_source, &bidi_spans);
     let manifest_path = rel_cwd_manifest_path(manifest_path, gctx);
     let mut emitted_source = None;
     for event in events {
@@ -99,13 +101,52 @@ pub fn text_direction_codepoint_in_comment(
         let token_span = token_span.start()..token_span.end();
         let mut snippet = Snippet::source(contents).path(&manifest_path).annotation(
             AnnotationKind::Context
-                .span(token_span)
-                .label("this comment contains an invisible unicode text flow control codepoint"),
+                .span(token_span.clone())
+                .label("this literal contains an invisible unicode text flow control codepoint"),
         );
         for bidi_span in event.bidi_spans {
             let bidi_span = bidi_span.0..bidi_span.1;
             let escaped = format!("{:?}", &contents[bidi_span.clone()]);
             snippet = snippet.annotation(AnnotationKind::Primary.span(bidi_span).label(escaped));
+        }
+        let mut help_snippet = Snippet::source(contents).path(&manifest_path);
+        if let Some(original_raw) = toml_source.get(&event.token) {
+            let mut decoded = String::new();
+            let replacement = match event.token.kind() {
+                toml_parser::parser::EventKind::SimpleKey => {
+                    use toml_writer::ToTomlKey as _;
+                    original_raw.decode_key(&mut decoded, &mut ());
+                    let builder = toml_writer::TomlKeyBuilder::new(&decoded);
+                    let replacement = builder.as_basic();
+                    Some(replacement.to_toml_key())
+                }
+                toml_parser::parser::EventKind::Scalar => {
+                    use toml_writer::ToTomlValue as _;
+                    let kind = original_raw.decode_scalar(&mut decoded, &mut ());
+                    if matches!(kind, toml_parser::decoder::ScalarKind::String) {
+                        let builder = toml_writer::TomlStringBuilder::new(&decoded);
+                        let replacement = match event.token.encoding() {
+                            Some(toml_parser::decoder::Encoding::BasicString)
+                            | Some(toml_parser::decoder::Encoding::LiteralString)
+                            | None => builder.as_basic(),
+                            Some(toml_parser::decoder::Encoding::MlBasicString)
+                            | Some(toml_parser::decoder::Encoding::MlLiteralString) => {
+                                builder.as_ml_basic()
+                            }
+                        };
+                        Some(replacement.to_toml_value())
+                    } else {
+                        None
+                    }
+                }
+                _ => None,
+            };
+            if let Some(mut replacement) = replacement {
+                for (bidi, escaped, _) in UNICODE_BIDI_CODEPOINTS {
+                    replacement = replacement.replace(*bidi, escaped);
+                }
+                help_snippet = help_snippet.patch(Patch::new(token_span.clone(), replacement));
+            }
         }
 
         let level = lint_level.to_diagnostic_level();
@@ -115,23 +156,25 @@ pub fn text_direction_codepoint_in_comment(
             primary = primary.element(Level::NOTE.message(emitted_source.as_ref().unwrap()));
         }
 
-        let report = [primary];
+        let help = Group::with_title(Level::HELP.secondary_title("if you want to keep them but make them visible in your source code, you can escape them")).element(help_snippet);
+
+        let report = [primary, help];
         gctx.shell().print_report(&report, lint_level.force())?;
     }
 
     Ok(())
 }
 
-const UNICODE_BIDI_CODEPOINTS: &[(char, &str)] = &[
-    ('\u{202A}', "LEFT-TO-RIGHT EMBEDDING"),
-    ('\u{202B}', "RIGHT-TO-LEFT EMBEDDING"),
-    ('\u{202C}', "POP DIRECTIONAL FORMATTING"),
-    ('\u{202D}', "LEFT-TO-RIGHT OVERRIDE"),
-    ('\u{202E}', "RIGHT-TO-LEFT OVERRIDE"),
-    ('\u{2066}', "LEFT-TO-RIGHT ISOLATE"),
-    ('\u{2067}', "RIGHT-TO-LEFT ISOLATE"),
-    ('\u{2068}', "FIRST STRONG ISOLATE"),
-    ('\u{2069}', "POP DIRECTIONAL ISOLATE"),
+const UNICODE_BIDI_CODEPOINTS: &[(char, &str, &str)] = &[
+    ('\u{202A}', r"\u{202A}", "LEFT-TO-RIGHT EMBEDDING"),
+    ('\u{202B}', r"\u{202B}", "RIGHT-TO-LEFT EMBEDDING"),
+    ('\u{202C}', r"\u{202C}", "POP DIRECTIONAL FORMATTING"),
+    ('\u{202D}', r"\u{202D}", "LEFT-TO-RIGHT OVERRIDE"),
+    ('\u{202E}', r"\u{202E}", "RIGHT-TO-LEFT OVERRIDE"),
+    ('\u{2066}', r"\u{2066}", "LEFT-TO-RIGHT ISOLATE"),
+    ('\u{2067}', r"\u{2067}", "RIGHT-TO-LEFT ISOLATE"),
+    ('\u{2068}', r"\u{2068}", "FIRST STRONG ISOLATE"),
+    ('\u{2069}', r"\u{2069}", "POP DIRECTIONAL ISOLATE"),
 ];
 
 struct BiDiEvent {
@@ -139,11 +182,10 @@ struct BiDiEvent {
     bidi_spans: Vec<(usize, usize)>,
 }
 
-fn bidi_events(contents: &str, bidi_spans: &[(usize, usize)]) -> Vec<BiDiEvent> {
+fn bidi_events(source: &Source<'_>, bidi_spans: &[(usize, usize)]) -> Vec<BiDiEvent> {
     let mut bidi_spans = bidi_spans.iter();
     let bidi_span = bidi_spans.next().copied();
 
-    let source = Source::new(contents);
     let tokens = source.lex().into_vec();
     let mut collector = BiDiCollector {
         bidi_span,
@@ -188,7 +230,20 @@ impl BiDiCollector<'_> {
 }
 
 impl EventReceiver for BiDiCollector<'_> {
-    fn comment(&mut self, span: Span, _error: &mut dyn toml_parser::ErrorSink) {
-        self.process(EventKind::Comment, None, span)
+    fn simple_key(
+        &mut self,
+        span: Span,
+        encoding: Option<Encoding>,
+        _error: &mut dyn toml_parser::ErrorSink,
+    ) {
+        self.process(EventKind::SimpleKey, encoding, span)
+    }
+    fn scalar(
+        &mut self,
+        span: Span,
+        encoding: Option<Encoding>,
+        _error: &mut dyn toml_parser::ErrorSink,
+    ) {
+        self.process(EventKind::Scalar, encoding, span)
     }
 }
